pyproject.toml
[project]
name = "roo_bricks"
version = "0.0.1"
readme = "ReadMe.md"
requires-python = ">=3.11.13"
dependencies = [
    "databricks-cli>=0.18.0",
    "databricks-connect>=15.1.3",
    "databricks-sdk[notebook]>=0.68.0",
    "duckdb>=1.4.1",
    "python-dotenv>=1.1.1",
]
